Update DW_decompo_slow_case.jl
We have two methods for the main iteration: 1. solve the full J vectors in parallel and then add all cuts; 2. solve a fraction of the models then see if cut is already generated and early break.

This code implements the 1st method.

Which one of the two method is better, is to be tested in large-scale case.

The 2nd method is https://discourse.julialang.org/t/julia-crashes-without-reporting-anything-when-i-optimize-a-vector-of-models-in-parallel/131801?u=waltermadelim